// web/src/lib/api/generated/models/internal-activitylog-activity-log-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * POS Kasir API
 * POS Kasir API
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { POSKasirInternalActivitylogRepositoryLogActionType } from './poskasir-internal-activitylog-repository-log-action-type';
// May contain unused imports in some cases
// @ts-ignore
import type { POSKasirInternalActivitylogRepositoryLogEntityType } from './poskasir-internal-activitylog-repository-log-entity-type';

export interface InternalActivitylogActivityLogResponse {
    'action_type'?: POSKasirInternalActivitylogRepositoryLogActionType;
    'created_at'?: string;
    'details'?: { [key: string]: any; };
    'entity_id'?: string;
    'entity_type'?: POSKasirInternalActivitylogRepositoryLogEntityType;
    'id'?: string;
    'user_id'?: string;
    'user_name'?: string;
}